+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[FIX]** `<Profile>`: name should be display in DISPLAY1 instead of DISPLAY2 when medium
 - [...]
 
 # v10.0.0 (07/08/2019)